requirements.txt
absl-py==1.4.0
aioftp==0.21.4
aiohttp==3.8.5
aiosignal==1.3.1
astropy==5.2.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.2.0
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
dm-tree==0.1.8
drms==0.6.4
executing==1.2.0
flatbuffers==23.5.26
fonttools==4.42.0
fpdf==1.7.2
frozenlist==1.4.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==2.8
importlib-metadata==6.7.0
importlib-resources==6.0.1
ipython==8.12.2
isodate==0.6.1
jedi==0.19.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parfive==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prettytable==3.8.0
prompt-toolkit==3.0.39
protobuf==4.23.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.25.0
pyerfa==2.0.0.3
Pygments==2.16.1
PyGObject==3.36.0
pyparsing==3.0.9
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rsa==4.9
scipy==1.10.1
six==1.14.0
soupsieve==2.4.1
stack-data==0.6.2
sunpy==4.1.7
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.21.0
termcolor==2.3.0
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.25.8
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
zeep==4.2.1
zipp==3.15.0
